refactor: extract browser agent functions to src/ui/browser_agent.py
- Extracted run_browser_agent() function (109 lines)
- Extracted chrome_restart_dialog() function
- Extracted show_restart_dialog() function (107 lines)
- Total reduction: 204 lines
- File size: 2988 → 2790 lines (-198 lines due to blank line cleanup)

Related to #326 Phase 3.2
